pyproject.toml
[project]
name = "claude-assistant"
version = "0.1.0"
description = "Claude personal assistant daemon"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.48,<0.2.0",
    "croniter>=6.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "discord.py>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ty>=0.0.1a0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
addopts = "-m 'not live_api and not integration'"
markers = [
    "live_api: tests that call real Claude API (run with -m live_api)",
    "integration: tests that require a running server (run with -m integration)",
]
# tests/live/ has its own conftest that removes the SDK mock and imports the
# real claude_agent_sdk.  Recursing into it during normal collection breaks
# all other tests that depend on the mock.  Run live tests explicitly:
#   uv run --group dev pytest tests/live/ -m live_api
norecursedirs = ["tests/live"]

[tool.ruff]
line-length = 100
indent-width = 2

[tool.ruff.lint]
select = ["E", "F", "I"]  # E=pycodestyle, F=pyflakes, I=isort (import sorting)
ignore = ["E501"]  # Line too long - let formatter handle it, don't error on string literals

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
indent-style = "space"

[tool.ty.environment]
# Extra paths for modules that are dynamically added to sys.path
extra-paths = [
    "skills/contacts/scripts",
    "skills/reminders/scripts",
    "skills/sms-assistant/scripts",
    "skills/tts/scripts",
]
